compiles and runs and visualizes
